requirements.txt
tensorflow==2.14.0
numpy==1.24.3
matplotlib==3.7.1
streamlit==1.28.2
pillow==10.2.0
h5py==3.9.0
